# Conflicts:
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/repository/ReportRepository.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/AccountingCoreResourceNetSuiteMock.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/MetricController.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/ReportController.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/requests/ReportSearchRequest.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/requests/metric/GetMetricDataRequest.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/views/ReportView.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/resource/views/metric/MetricView.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/internal/ReportService.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/internal/metrics/MetricService.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/internal/metrics/MetricServiceImpl.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/internal/metrics/executors/BalanceSheetMetricService.java
#	accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/internal/metrics/executors/IncomeStatementMetricService.java
#	accounting_reporting_core/src/test/java/org/cardanofoundation/lob/app/accounting_reporting_core/service/business_rules/items/AccountAccountEventCodesConversionTaskItemTest.java
#	blockchain_common/src/main/java/org/cardanofoundation/lob/app/blockchain_common/service_assistance/JsonSchemaMetadataChecker.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/config/TransactionSubmissionConfig.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/domain/core/BlockchainTransactions.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/domain/entity/txs/TransactionEntity.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/job/TransactionsWatchDogJob.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/repository/ReportEntityRepositoryGateway.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/repository/TransactionEntityRepositoryGateway.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/API1L1TransactionCreator.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/BlockchainPublishStatusMapper.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/ReportSerializer.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/TransactionsWatchDogService.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/dispatch/BlockchainReportsDispatcher.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/dispatch/BlockchainTransactionsDispatcher.java
#	blockchain_publisher/src/main/java/org/cardanofoundation/lob/app/blockchain_publisher/service/transation_submit/DefaultTransactionSubmissionService.java
#	blockchain_publisher/src/test/java/org/cardanofoundation/lob/app/blockchain_publisher/service/TransactionsWatchDogServiceTest.java
#	organisation/src/main/java/org/cardanofoundation/lob/app/organisation/service/OrganisationService.java
#	support/src/main/java/org/cardanofoundation/lob/app/support/spring_audit/internal/AuditDataProvider.java

+++ accounting_reporting_core/src/main/java/org/cardanofoundation/lob/app/accounting_reporting_core/domain/core/report/Report.java
@@ -30,7 +30,7 @@ public class Report {
 
     private short year;
 
-    private Integer ver;
+    private long ver;
 
     @Builder.Default
     private Optional<Short> period = Optional.empty();
@@ -59,7 +59,7 @@ public class Report {
                             ReportType reportType,
                             IntervalType intervalType,
                             short year,
-                            Integer ver,
+                            long ver,
                             Optional<Short> period) {
         return period.map(p -> {
             return digestAsHex(STR."\{organisationId}::\{reportType}::\{intervalType}::\{year}::\{ver}::\{p}");
